feat(vehicles): add price per gallon chart to vehicle detail
- PriceHistoryChart shows line chart of fuel price history using Charts
  framework, with min/avg/high summary labels below
- Chart only appears when 2+ gas fill-ups have price data
- Fix VehicleEvent.hours from Int? to Double? to match backend DecimalField

+++ Sources/HaleHubIOS/Core/Models/Vehicle.swift
@@ -39,7 +39,7 @@ struct VehicleEvent: Identifiable, Codable, Sendable {
     let eventType: String
     let date: Date
     let miles: Int?
-    let hours: Int?
+    let hours: Double?
     let gallons: Double?
     let pricePerGallon: Double?
     let totalCost: Double?
